@@ -10,14 +10,14 @@ description: Teleblitz app
 version: 1.1.0+10
 
 environment:
-  sdk: ^3.1.1
+  sdk: '>=3.3.0 <4.0.0'
+  # ^3.1.1
 
 dependencies:
   flutter_localizations:
     sdk: flutter
   flutter:
     sdk: flutter
-  chewie: ^1.0.0
   flutter_html: any
   http: any
   intl: any
@@ -29,9 +29,8 @@ dependencies:
   rxdart: ^0.27.7
   flip_card: ^0.7.0
   showcaseview: ^2.0.3
-  device_info_plus: ^9.1.2
+  device_info_plus: ^10.0.1
   shared_preferences: any
-  google_maps_place_picker_mb: ^3.1.2
   flutter_masked_text2: ^0.9.1
   platform: ^3.1.0
   cupertino_icons: ^1.0.2
@@ -42,7 +41,6 @@ dependencies:
   firebase_messaging: ^14.7.19
   flutter_datetime_picker_plus: ^2.1.0
   flutter_barcode_scanner: ^2.0.0
-  wakelock_plus: ^1.2.1
 
 dev_dependencies:
   flutter_launcher_icons: "^0.13.1"
