@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:morea/Widgets/standart/buttons.dart';
 import 'package:morea/Widgets/standart/moreaTextStyle.dart';
+import 'package:morea/morea_strings.dart';
 import 'package:morea/morealayout.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:morea/services/utilities/moreaInputValidator.dart';
-import 'package:google_maps_place_picker_mb/google_maps_place_picker.dart';
 
 class ChangeAbtreten extends StatefulWidget {
   final String antreten, mapAntreten;
@@ -163,7 +162,110 @@ class _ChangeAbtretenState extends State<ChangeAbtreten> {
                         ),
                         Padding(
                             padding: const EdgeInsets.only(top: 10.0),
-                            child: moreaRaisedButton(this.nameMapAbtreten, () {
+                            child: Column(
+                              children: [
+                                TextButton(
+                                  child: Container(
+                                    constraints: BoxConstraints(
+                                        minWidth: viewportConstraints.maxWidth,
+                                        maxWidth: viewportConstraints.maxWidth),
+                                    child: Padding(
+                                      padding: const EdgeInsets.all(8.0),
+                                      child: Text('Lokal',
+                                        style: MoreaTextStyle.textField,
+                                      ),
+                                    ),
+                                  ),
+                                  style: ButtonStyle(
+                                      shape: MaterialStateProperty.all<OutlinedBorder>(
+                                          RoundedRectangleBorder(
+                                              borderRadius: BorderRadius.circular(5),
+                                              side: BorderSide(color: Colors.black45))),
+                                      foregroundColor: MaterialStateProperty.all<Color>(
+                                          Colors.black),
+                                      padding: MaterialStateProperty.all<EdgeInsets>(
+                                          EdgeInsets.only(
+                                              top: 10, bottom: 10, left: 5, right: 5)),
+                                      overlayColor: MaterialStateProperty.resolveWith(
+                                              (Set<MaterialState> states) {
+                                            if (states.contains(MaterialState.focused))
+                                              return MoreaColors.violett;
+                                            return null;
+                                          })),
+                                  onPressed: () {
+                                    this.urlMapAbtreten = urlGoogleMapsLokal;
+                                    this.nameMapAbtreten = 'Lokal';
+                                  },
+                                ),
+                                TextButton(
+                                  child: Container(
+                                    constraints: BoxConstraints(
+                                        minWidth: viewportConstraints.maxWidth,
+                                        maxWidth: viewportConstraints.maxWidth),
+                                    child: Padding(
+                                      padding: const EdgeInsets.all(8.0),
+                                      child: Text('Waldrand',
+                                        style: MoreaTextStyle.textField,
+                                      ),
+                                    ),
+                                  ),
+                                  style: ButtonStyle(
+                                      shape: MaterialStateProperty.all<OutlinedBorder>(
+                                          RoundedRectangleBorder(
+                                              borderRadius: BorderRadius.circular(5),
+                                              side: BorderSide(color: Colors.black45))),
+                                      foregroundColor: MaterialStateProperty.all<Color>(
+                                          Colors.black),
+                                      padding: MaterialStateProperty.all<EdgeInsets>(
+                                          EdgeInsets.only(
+                                              top: 10, bottom: 10, left: 5, right: 5)),
+                                      overlayColor: MaterialStateProperty.resolveWith(
+                                              (Set<MaterialState> states) {
+                                            if (states.contains(MaterialState.focused))
+                                              return MoreaColors.violett;
+                                            return null;
+                                          })),
+                                  onPressed: () {
+                                    this.urlMapAbtreten = urlGoogleMapsWaldrand;
+                                    this.nameMapAbtreten = 'Waldrand';
+                                  },
+                                ),
+                                TextButton(
+                                  child: Container(
+                                    constraints: BoxConstraints(
+                                        minWidth: viewportConstraints.maxWidth,
+                                        maxWidth: viewportConstraints.maxWidth),
+                                    child: Padding(
+                                      padding: const EdgeInsets.all(8.0),
+                                      child: Text('Alle anderen Orte',
+                                        style: MoreaTextStyle.textField,
+                                      ),
+                                    ),
+                                  ),
+                                  style: ButtonStyle(
+                                      shape: MaterialStateProperty.all<OutlinedBorder>(
+                                          RoundedRectangleBorder(
+                                              borderRadius: BorderRadius.circular(5),
+                                              side: BorderSide(color: Colors.black45))),
+                                      foregroundColor: MaterialStateProperty.all<Color>(
+                                          Colors.black),
+                                      padding: MaterialStateProperty.all<EdgeInsets>(
+                                          EdgeInsets.only(
+                                              top: 10, bottom: 10, left: 5, right: 5)),
+                                      overlayColor: MaterialStateProperty.resolveWith(
+                                              (Set<MaterialState> states) {
+                                            if (states.contains(MaterialState.focused))
+                                              return MoreaColors.violett;
+                                            return null;
+                                          })),
+                                  onPressed: () {
+                                    this.urlMapAbtreten = urlGoogleMapsEmpty;
+                                    this.nameMapAbtreten = 'Alle anderen Orte';
+                                  },
+                                ),
+                              ],
+                            )
+                            /*moreaRaisedButton(this.nameMapAbtreten, () {
                               Navigator.push(
                                   context,
                                   MaterialPageRoute(
@@ -183,7 +285,8 @@ class _ChangeAbtretenState extends State<ChangeAbtreten> {
                                             usePlaceDetailSearch: true,
                                             useCurrentLocation: false,
                                           )));
-                            })),
+                            })*/
+                        ),
                       ],
                     ),
                   ),
